Leave Automation Web App/requirements.txt
selenium==4.6.0
requests==2.28.1
webdriver_manager==3.8.4
packaging==21.3
flask-restful==0.3.9
gunicorn==20.1.0
uvicorn==0.30.6
